refactor: Migrate TrainerPersona context to hooks structure
- Refactored TrainerPersonaContext to utilize a centralized hooks structure, improving organization and maintainability.
- Updated imports in various components to use hooks from the new hooks directory, enhancing clarity and reducing redundancy.
- Removed obsolete context definitions and custom hooks related to TrainerPersona, streamlining the codebase for better performance and readability.
